Various bug fixes.

+++ src/Merit/Merit.Web/Pages/EditCompanyWants.cshtml.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Merit.AdvertisementService;
 using Merit.Data.Models;
 using Merit.WantsService;
 using Microsoft.AspNetCore.Identity;
@@ -13,6 +14,7 @@ namespace Merit.Web.Pages
     public class EditCompanyWantsModel : PageModel
     {
         private IWantsService wService = new WantsService.WantsService();
+        private IAdvertisementService advertisementService = new AdvertisementService.AdvertisementService();
 
         private readonly UserManager<IdentityUser> userManager;
         private readonly SignInManager<IdentityUser> signInManager;
@@ -23,7 +25,7 @@ namespace Merit.Web.Pages
             this.userManager = userManager;
         }
 
-        [BindProperty(SupportsGet = true)]
+        [BindProperty]
         public List<CompanyWants> CompanyWantsList { get; set; }
 
         [BindProperty(SupportsGet = true)]
@@ -37,8 +39,11 @@ namespace Merit.Web.Pages
 
         public string Message { get; set; }
 
-        [BindProperty(SupportsGet = true)]
+        [BindProperty]
         public bool Visi { get; set; } = false;
+
+        [BindProperty]
+        public bool Success { get; set; }
 
         public async Task<IActionResult> OnGetAsync()
         {
@@ -53,6 +58,10 @@ namespace Merit.Web.Pages
             if (cUser is CompanyUser companyUser)
             {
                 CompanyWantsList = wService.GetAllCompanyWants(companyUser.CompanyUserId);
+                if (SelectedCompanyWantId != 0 && CompanyWantsList.FirstOrDefault(x => x.CompanyWantsId == SelectedCompanyWantId) == null)
+                {
+                    return Redirect($"/CompanyInfoPage");
+                }
             }
             else if (cUser is PersonalUser)
             {
@@ -60,20 +69,29 @@ namespace Merit.Web.Pages
             }
             foreach (var want in CompanyWantsList)
             {
-                if(want.CompanyWantsId == SelectedCompanyWantId)
+                if (want.CompanyWantsId == SelectedCompanyWantId)
                 {
                     SelectedCompanyWantId = want.CompanyWantsId;
                     WantText = want.Want;
                 }
 
             }
+
+            if (SelectedCompanyWantId != 0 && advertisementService.WantExistsInAdvertisement(SelectedCompanyWantId))
+            {
+                Success = false;
+                Visi = true;
+                Message = "Obs! Denna want är länkad till en eller flera annonser.";
+            }
             return Page();
         }
 
         public async Task OnPostEdit()
         {
+
             wService.EditCompanyWant(CWant);
             Visi = true;
+            Success = true;
             Message = "Önskemål ändrat";
             SelectedCompanyWantId = 0;
             await OnGetAsync();
@@ -81,11 +99,24 @@ namespace Merit.Web.Pages
 
         public async Task OnPostDelete()
         {
-            wService.DeleteCompanyWant(CWant);
-            Visi = true;
-            Message = "Önskemål borttaget";
-            SelectedCompanyWantId = 0;
+
+            if (advertisementService.WantExistsInAdvertisement(CWant.CompanyWantsId))
+            {
+                Message = "Du kan inte ta bort wanten, den används i en annons.";
+                Visi = true;
+                Success = false;
+                SelectedCompanyWantId = 0;
+            }
+            else
+            {
+                wService.DeleteCompanyWant(CWant);
+                Visi = true;
+                Success = true;
+                Message = "Önskemål borttaget";
+                SelectedCompanyWantId = 0;
+            }
+
             await OnGetAsync();
-        } 
+        }
     }
 }
